+++ frontend/src/app/main/ui/shapes/attrs.cljs
@@ -113,9 +113,9 @@
 
      (obj/merge! attrs (clj->js fill-attrs)))))
 
-(defn add-stroke [attrs shape render-id]
+(defn add-stroke [attrs shape render-id index]
   (let [stroke-style (:stroke-style shape :none)
-        stroke-color-gradient-id (str "stroke-color-gradient_" render-id)
+        stroke-color-gradient-id (str "stroke-color-gradient_" render-id "_" index)
         stroke-width (:stroke-width shape 1)]
     (if (not= stroke-style :none)
       (let [stroke-attrs
@@ -198,14 +198,13 @@
 
          styles (-> (obj/get props "style" (obj/new))
                     (obj/merge! svg-styles)
-                    (add-stroke shape render-id)
                     (add-layer-props shape))
 
          styles (cond (or (some? (:fill-image shape))
                           (= :image (:type shape))
                           (> (count (:fills shape)) 1)
                           (some #(some? (:fill-color-gradient %)) (:fills shape)))
-                  (obj/set! styles "fill" (str "url(#fill-0-" render-id ")"))
+                      (obj/set! styles "fill" (str "url(#fill-0-" render-id ")"))
 
                       ;; imported svgs can have fill and fill-opacity attributes
                       (obj/contains? svg-styles "fill")
@@ -233,7 +232,15 @@
     (-> (obj/new)
         (obj/set! "style" fill-styles))))
 
+(defn extract-stroke-attrs
+  [shape index]
+  (let [render-id (mf/use-ctx muc/render-ctx)
+        stroke-styles (-> (obj/get shape "style" (obj/new))
+                          (add-stroke shape render-id index))]
+    (-> (obj/new)
+        (obj/set! "style" stroke-styles))))
+
 (defn extract-border-radius-attrs
   [shape]
-   (-> (obj/new)
-       (add-border-radius shape)))
+  (-> (obj/new)
+      (add-border-radius shape)))