+++ frontend/src/app/main/ui/shapes/image.cljs
@@ -8,7 +8,7 @@
   (:require
    [app.common.geom.shapes :as gsh]
    [app.main.ui.shapes.attrs :as attrs]
-   [app.main.ui.shapes.custom-stroke :refer [shape-custom-stroke]]
+   [app.main.ui.shapes.custom-stroke :refer [shape-custom-strokes]]
    [app.util.object :as obj]
    [rumext.alpha :as mf]))
 
@@ -29,8 +29,7 @@
                         :height height}))
         path? (some? (.-d props))]
 
-    [:g
-     [:& shape-custom-stroke {:shape shape}
-      (if path?
-        [:> :path props]
-        [:> :rect props])]]))
+    [:& shape-custom-strokes {:shape shape}
+     (if path?
+       [:> :path props]
+       [:> :rect props])]))