+++ frontend/src/app/main/ui/workspace/colorpalette.cljs
@@ -45,7 +45,7 @@
         select-color
         (fn [event]
           (if (kbd/alt? event)
-              (st/emit! (mdc/change-stroke ids-with-children (merge uc/empty-color color)))
+              (st/emit! (mdc/change-stroke ids-with-children (merge uc/empty-color color) 0))
               (st/emit! (mdc/change-fill ids-with-children (merge uc/empty-color color) 0))))]
 
     [:div.color-cell {:on-click select-color}